minor revision

+++ examples/battle_model/algo/dgn.py
@@ -1,3 +1,4 @@
+import numpy as np
 import os
 import random
 import torch
@@ -69,7 +70,7 @@ class DGNAgent(nn.Module):
             q_value = self.q_net(graph)
             action = q_value.argmax(dim=-1).detach()
 
-        return action.numpy().astype(int)
+        return action.numpy().astype(np.int32)
 
     def get_q(self, graph):
         z, weight = self.conv_net(graph)
